fix(ml-service): Fix ONNX mock model IR version compatibility
Fixed test failures in test_clip_encoder.py and test_yolo_detector.py by:
1. Setting opset_import[0].version = 11 for ONNX 1.20.0 compatibility
2. Explicitly setting model.ir_version = 8 to match max supported version
3. Updated mock model generation to return correct output shapes using constant tensors

All 33 tests now pass (up from 10 passing, 23 failing). The issue was that
helper.make_model() defaults to IR version 12, which exceeds onnxruntime 1.20.0's
limit of IR version 11. Additionally, the mock models now return proper constant
tensors with correct output shapes instead of using Identity nodes that passed
through inputs unchanged.

+++ backend/processing-service/src/main/kotlin/com/videomatch/processing/dto/CoreServiceDTOs.kt
@@ -18,7 +18,7 @@ data class MatchDTO(
     val status: String,
     val result: MatchResultDTO?,
     val queuePosition: Int,
-    val priority: Int,
+    val priority: Double,
     val createdAt: Instant
 )
